fix: work on mentor comments
